nrfx: Use actual nrfx instead of manually modified files
With nrfx 2.3.0 it is possible to include the MDK files on
non-ARM platforms. As the MDK has a depenency to cmsis core
files, a simple mocked core_cm4.h has been provided.

By defining NRF_DECLARE_ONLY and NRF_STATIC_INLINE to nothing,
no functions are inlined. This ensures that it is possible to
mock the peripherals.

For a non-zephyr build, the application must provide the
environment varianble NRFX_BASE.

The user of the nRF MDK or nrfx must ensure that the peripheral
pointers are redefined before they are used. This is done in
nrfx_bsim_redef.h. Include this header after including nrf.h
or similar. When using nrfx, this inclusion can be placed in
nrfx_glue.h

For non-zephyr builds, a nrfx_config and nrfx_glue is provided.

The original implementation of the HAL APIs from the real nrfx
is included in `nrf_hal_originals.c`.
All these functions are labelled as weak,
so they can be replaced as needed with new definition in `src/nrfx/hal/`
This removes the need for duplicating code.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>